Moved HyphenateHttpHeadersRule from .zalando to .zally package
